Check permissions when working with notes

+++ note_14/database/notes.py
@@ -10,40 +10,42 @@ get_notes: Get all the notes from the database that a user can view
 
 from datetime import date
 
+from sqlalchemy.orm.exc import NoResultFound
+
 from config import PermissionType
-from models import Note, NoteSection
-
-from database import add_permission
-
-
-class UnauthorizedError(Exception):
-    """Raised when a user attempts an action they are not authorized for."""
+from models import Note, NotePermission, NoteSection
+from database import UnauthorizedError, add_permission, has_permission
 
 
 def get_notes(session, user):  # pylint: disable=unused-argument
     """ Get all notes from the database that a user can see. """
 
-    notes = user.notes
-
-    # TODO: Add notes that user has view permission for
+    notes = [
+        permission.note
+        for permission in user.permissions
+        if permission.type == PermissionType.READ
+    ]
 
     return notes
 
 
-def get_note(session, note_id, user):  # pylint: disable=unused-argument
+def get_note(session, note_id, user):
     """Get a single note from the database.
 
     Will report an error if the user is not authorized to view that note.
     """
 
-    # TODO: Do a more explicit check using permissions once they are set up
-    for note in user.notes:
-        if note.id == note_id:
-            return note
-
-    raise UnauthorizedError(
-        f"{user} not authorized to access {note_id} or it does not exist"
-    )
+    try:
+        return (
+            session.query(NotePermission)
+            .filter_by(user_id=user.id, note_id=note_id, type=PermissionType.READ)
+            .one()
+            .note
+        )
+    except NoResultFound as error:
+        raise UnauthorizedError(
+            f"{user} not authorized to access {note_id} or it does not exist"
+        ) from error
 
 
 def create_note(session, title, text, user):
@@ -67,11 +69,19 @@ def create_note(session, title, text, user):
 def delete_note(session, note_id, user):
     """ Delete a note from the database """
     note = get_note(session, note_id, user)
+
+    if not has_permission(session, PermissionType.EDIT, user, note):
+        raise UnauthorizedError(f"{user} not authorized to delete {note}")
+
     session.delete(note)
 
 
 def edit_note(session, title, text, note_id, user):
     """ edit an existing note in the database. """
     note = get_note(session, note_id, user)
+
+    if not has_permission(session, PermissionType.EDIT, user, note):
+        raise UnauthorizedError(f"{user} not authorized to delete {note}")
+
     note.title = title
     note.text = text
